# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pinjie Fullstack Blog Backend",
    "description": "独立个人博客的后端 API，当前提供认证、账户、后台管理、文件资产和系统设置基础能力。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/settings/site": {
      "get": {
        "tags": [
          "系统设置"
        ],
        "summary": "获取站点设置",
        "operationId": "get_site_setting_api_v1_admin_settings_site_get",
        "parameters": [
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_AdminSiteSettingRead_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "系统设置"
        ],
        "summary": "更新站点设置",
        "operationId": "update_site_setting_api_v1_admin_settings_site_patch",
        "parameters": [
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteSettingPatchIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_AdminSiteSettingRead_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/settings/site/logo": {
      "put": {
        "tags": [
          "系统设置"
        ],
        "summary": "更换站点 LOGO",
        "operationId": "update_site_logo_api_v1_admin_settings_site_logo_put",
        "parameters": [
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_update_site_logo_api_v1_admin_settings_site_logo_put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_AdminSiteSettingRead_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "系统设置"
        ],
        "summary": "移除站点 LOGO",
        "operationId": "delete_site_logo_api_v1_admin_settings_site_logo_delete",
        "parameters": [
          {
            "name": "revision",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "description": "读取站点设置时获得的修订号",
              "title": "Revision"
            },
            "description": "读取站点设置时获得的修订号"
          },
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_AdminSiteSettingRead_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/settings/registration": {
      "get": {
        "tags": [
          "系统设置"
        ],
        "summary": "获取注册设置",
        "operationId": "get_registration_setting_api_v1_admin_settings_registration_get",
        "parameters": [
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_AdminRegistrationSettingRead_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "系统设置"
        ],
        "summary": "更新注册设置",
        "operationId": "update_registration_setting_api_v1_admin_settings_registration_patch",
        "parameters": [
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationSettingPatchIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_AdminRegistrationSettingRead_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminRegistrationSettingRead": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "是否允许 Web 公开注册普通用户"
          },
          "revision": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Revision",
            "description": "当前设置修订号"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "最近更新时间"
          },
          "updated_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdminSummaryRead"
              },
              {
                "type": "null"
              }
            ],
            "description": "最后修改管理员摘要"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "enabled",
          "revision",
          "updated_at",
          "updated_by"
        ],
        "title": "AdminRegistrationSettingRead"
      },
...
      "AdminSiteSettingRead": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Web 公共站点名称"
          },
          "logo": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SiteLogoRead"
              },
              {
                "type": "null"
              }
            ],
            "description": "站点 LOGO 公开信息"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Web 默认页面标题"
          },
          "keywords": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Keywords",
            "description": "Web Metadata 关键词"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Web 默认站点描述"
          },
          "revision": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Revision",
            "description": "当前设置修订号"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "最近更新时间"
          },
          "updated_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdminSummaryRead"
              },
              {
                "type": "null"
              }
            ],
            "description": "最后修改管理员摘要"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "logo",
          "title",
          "keywords",
          "description",
          "revision",
          "updated_at",
          "updated_by"
        ],
        "title": "AdminSiteSettingRead"
      },
      "AdminSummaryRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "最后修改管理员唯一标识"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "最后修改管理员显示名称"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "display_name"
        ],
        "title": "AdminSummaryRead"
      },
...
      "Body_update_site_logo_api_v1_admin_settings_site_logo_put": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "PNG、JPEG 或 WebP 静态 LOGO，最大 2 MB"
          },
          "revision": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Revision",
            "description": "读取站点设置时获得的修订号"
          }
        },
        "type": "object",
        "required": [
          "file",
          "revision"
        ],
        "title": "Body_update_site_logo_api_v1_admin_settings_site_logo_put"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail",
            "description": "请求参数校验错误详情列表"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegistrationSettingPatchIn": {
        "properties": {
          "revision": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Revision",
            "description": "读取设置时获得的修订号"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "是否允许 Web 公开注册普通用户"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision",
          "enabled"
        ],
        "title": "RegistrationSettingPatchIn"
      },
...
      "ResponseModel_AdminRegistrationSettingRead_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "稳定程序代码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "面向调用方的中文结果消息"
          },
          "data": {
            "$ref": "#/components/schemas/AdminRegistrationSettingRead",
            "description": "响应业务数据"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "用于定位本次请求的唯一标识"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[AdminRegistrationSettingRead]"
      },
      "ResponseModel_AdminSiteSettingRead_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "稳定程序代码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "面向调用方的中文结果消息"
          },
          "data": {
            "$ref": "#/components/schemas/AdminSiteSettingRead",
            "description": "响应业务数据"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "用于定位本次请求的唯一标识"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[AdminSiteSettingRead]"
      },
...
      "SiteLogoRead": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "带修订号缓存参数的站点 LOGO 公开路径"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type",
            "description": "站点 LOGO 的真实 MIME 类型"
          },
          "file_size": {
            "type": "integer",
            "title": "File Size",
            "description": "站点 LOGO 文件大小，单位为字节"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "url",
          "mime_type",
          "file_size"
        ],
        "title": "SiteLogoRead"
      },
...
      "SiteSettingPatchIn": {
        "properties": {
          "revision": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Revision",
            "description": "读取设置时获得的修订号"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Web 公共站点名称"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Web 默认页面标题"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords",
            "description": "Web Metadata 关键词"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Web 默认站点描述"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "revision"
        ],
        "title": "SiteSettingPatchIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location",
            "description": "错误字段在请求中的位置"
          },
          "msg": {
            "type": "string",
            "title": "Message",
            "description": "参数校验错误消息"
          },
          "type": {
            "type": "string",
            "title": "Error Type",
            "description": "参数校验错误类型"
          },
          "input": {
            "title": "Input",
            "description": "引发校验错误的输入值，敏感内容可能被省略"
          },
          "ctx": {
            "type": "object",
            "title": "Context",
            "description": "参数校验错误的补充上下文"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}